chore(upgrade): improve effieciency of application - upgrade to use postgresql database - use serializers for query items and todos - refactored authentication
