+++ app/auth/authenticate_user.rb
@@ -14,7 +14,7 @@ class AuthenticateUser
 
   def user
     user = User.find_by(email: email)
-    return user if user && user.authenticate(password)
+    return user if user&.authenticate(password)
     raise(ExceptionHandler::AuthenticationError, Message.invalid_credentials)
   end
 end
