+++ app/models/user.rb
@@ -5,4 +5,5 @@ class User < ApplicationRecord
   has_many :todos, foreign_key: :created_by
 
   validates_presence_of :name, :email, :password_digest
+  validates_uniqueness_of :email
 end
