+++ app/serializers/todo_serializer.rb
@@ -1,0 +1,4 @@
+class TodoSerializer < ActiveModel::Serializer
+  attributes :id, :title, :created_by, :created_at, :updated_at
+  has_many :items
+end
