+++ config/routes.rb
@@ -1,7 +1,7 @@
 Rails.application.routes.draw do
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
-  post 'auth/login', to: 'authentication#authenticate'
-  post 'signup', to: 'users#create'
+  post 'auth/login', to: 'users#login'
+  post 'signup', to: 'users#signup'
   resources :todos do
     resources :items
   end
