+++ db/migrate/20190603224946_change_column_default.rb
@@ -1,0 +1,5 @@
+class ChangeColumnDefault < ActiveRecord::Migration[5.2]
+  def change
+    change_column_default :items, :done, false
+  end
+end
