+++ db/schema.rb
@@ -10,11 +10,14 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2019_03_08_162619) do
+ActiveRecord::Schema.define(version: 2019_06_03_224946) do
+
+  # These are extensions that must be enabled in order to support this database
+  enable_extension "plpgsql"
 
   create_table "items", force: :cascade do |t|
     t.string "name"
-    t.boolean "done"
+    t.boolean "done", default: false
     t.integer "todo_id"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
