+++ spec/requests/users_spec.rb
@@ -1,10 +1,22 @@
 require 'rails_helper'
 
 RSpec.describe 'Users API', type: :request do
-  let(:user) { build(:user) }
+  let!(:user) { create(:user) }
   let(:headers) { valid_headers.except('Authorization') }
   let(:valid_attributes) do
     attributes_for(:user, password_confirmation: user.password)
+  end
+  let(:valid_credentials) do
+    {
+        email: user.email,
+        password: user.password
+    }.to_json
+  end
+  let(:invalid_credentials) do
+    {
+        email: Faker::Internet.email,
+        password: Faker::Internet.password
+    }.to_json
   end
 
   describe 'POST /signup' do
@@ -34,6 +46,24 @@ RSpec.describe 'Users API', type: :request do
       it 'returns failure message' do
         expect(json['message'])
           .to match(/Validation failed: Password can't be blank, Name can't be blank, Email can't be blank, Password digest can't be blank/)
+      end
+    end
+  end
+
+  describe "POST /login" do
+    context 'When request is valid' do
+      before { post '/auth/login', params: valid_credentials, headers: headers }
+
+      it 'returns an authentication token' do
+        expect(json['auth_token']).not_to be_nil
+      end
+    end
+
+    context 'When request is invalid' do
+      before { post '/auth/login', params: invalid_credentials, headers: headers }
+
+      it 'returns a failure message' do
+        expect(json['message']).to match(/Invalid credentials/)
       end
     end
   end
